requirements.txt
fastapi
pydantic
gitpython
faiss-cpu
openai